[Operator] Remove dependency on istio CRDs
The operator had a hard dependency on istio CRDs to be present in
the cluster. That was the case, since the GerritCluster reconciler
would always check whether to reconcile istio resources, since a
different ingress provider could be used for every GerritCluster.

Now, the ingress provider to be used has to be configured on operator
level, i.e. the operator will only support a single ingress provider
at a time. It is anyway unlikely that within a single cluster an
ingress controller and istio would be used at the same time.

Depending on which ingress provider is being used, the operator will
register a different implementation of a reconciler that will reconcile
with the new GerritNetwork CustomResource. This CustomResource is not
meant to be manually applied, but will be applied by the operator
based on the GerritCluster Customresource.

Supported ingress providers are:

- NONE: no ingress provider will be installed. The user will have to set
  this up themselves. Optionally, the Gerrit configuration for ingress
  traffic can still be managed by the operator.
- INGRESS: the nginx ingress controller is used. SSH support is limited.
- ISTIO: istio is used to provide an ingress. SSH is fully supported.

Change-Id: I55be6c7c71d07e436e8c1cc6c4ddd5d69b78d652