[Operator] Route SSH requests by port
The routing of SSH requests with multiple Gerrit instances was not
working as intended. The subdomain was supposed to be used, but the
VirtualService does not provide matchers for that purpose.

There are two solutions, use the port instead or deploy a gateway for
each Gerrit. To keep the number of managed resources small, the
option of using ports was chosen here.

Change-Id: I0e7832e776498c1d428a46db123f969d5acb17cf