Fix cache for download plugins
To enable the cache for downloaded plugins the gerrit-initializer
used by the gerrit-init init container had to be configured with
the `pluginCache` option being set to `true`. That was a bad naming
choice for a boolean option. Further, the operator did set the
`pluginCacheEnabled` option instead, which didn't exist.

This change adds the `pluginCacheEnabled`-option to the gerrit-
initializer. It keeps the old option for backwards compatibility
but deprecates it.

Change-Id: I83aac715308d114fd14b8818f69eae29eb2407ff