[Operator] Merge volume storing logs into shared volume
As previously with the plugin cache this change merges the volume
for storing logs into the shared volume to reduce the number of
volumes managed by the operator.

This is a breaking change, since old logs won't be available anymore
from the pods and the volume storing the logs changes.

Change-Id: Ib8cc4c0472839e5846e2e46ae44ba3606f6eff37