[Operator] Install global-refdb module for scaled primary Gerrits
The high-availability plugin has a dependency to the global-refdb
lib module. Thus, as a preparing change to enable high availability
for primary Gerrits, this change automatically installs the
global-refdb module, if a highly-available primary Gerrit is being
installed.

To achieve that, the python scripts performing the initialization
were extended to automatically install required lib modules.

Change-Id: Icb9acaa755775c54386b299fbec78c7bb88e1d2b