[Operator] Support usage of a global ref-database
Two run two primary Gerrits in active-active mode, a global refdb
is required to avoid split brain scenarios.

This change adds the possibility to configure a global refdb for
a GerritCluster. Currently, only the zookeeper-refdb implementation
is supported, but others will be added in the future. If a high-
availability setup is being installed using the operator, usage
of a global refdb is required.

Change-Id: I1db88291ada791e8def63443ae4b750c11a8cb7d