[Operator] Use same serverId in all Gerrits in GerritCluster
In a GerritCluster all Gerrit instances serve the same repositories.
Thus, they need to have the same serverId. This was not ensured.

Now, the serverId can be set in the GerritCluster CustomResource.
If it is not set, the serverId will be set to the Namespace and
name of the GerritCluster. If Gerrits are installed separately,
i.e. not as part of a GerritCluster, it has to be set in the
gerrit.config.

Change-Id: I8067f70fe757d0d7d47655d6de50e51beb37825c