Don't shade repositories with apache-git-http-backend
The apache-git-http-backend / receiver used the /new and the /git
paths. All requests using those prefixes were routed to the receiver.
This meant repositories using the same prefix could not be cloned/
fetched from the Gerrit Replica.

Now, the routing is determined by using the two requests that are
part of a push, i.e. if the request is one of
  a) /repo/info/refs/?service=git-receive-pack
  b) /repo/git-receive-pack
it will be routed to the receiver. The same is still true for paths
starting with /a/projects/ since they are being used to replicate
repo creation etc. All other requests will be routed to the Gerrit
Replica.

As part of this change, the /new/ path to create new repositories,
which was deprecated some time ago was removed.

Further, the health check had to be adapted, since the path `/` now
requires a valid user. Instead of a HTTP GET request, now only a TCP
Socket check is done, i.e. a check whether the application listens on
the configured port. A proper health endpoint should be added in the
future.

Change-Id: Id5a88263589263ace12bc6b5bc7084917a6130da