Copy Build-Approved label
This label determines whether a change will be tested in the
CI. Only project owners are allowed to vote on it and should
only do so, if they deem it safe to deploy the setup in a
Kubernetes cluster.

To reduce the effort a little bit, assume that changes not
touching any code or uploaded by a project owner are safe
enough to copy the votes.

Change-Id: Iab1222a13fd53525b5e9192ea20bb34af8ef4817