Forbid outside access to ha-plugin REST API in istio
The high-availability plugin uses REST API endpoints to forward
messages, i.e. to trigger reindexing. These endpoints were accessible
from outside the cluster, which was insecure, if the plugin wasn't
configured to use authentication.

Since access from outside the cluster is not really required, so
this change just forbids calling the respective URI paths via
the gateway.

The canonical way to enforce such a policy would be to use an
AuthorizationPolicy. However, this only works properly when using mTLS
which is currently not set up when communicating between single
pods of a StatefulSet, which is done by the HA-plugin. Thus,
a workaround is used that makes use of the fault injection feature.
Requests calling (/a)?/plugins/high-availability/* will thus return
403 when accessed through the VirtualService.

Change-Id: I9e288e3282cdca56f418439f6415f15c687874b2