Allow to select istio ingressgateway
The label selector that was used by the Gateway CR to select the
Ingress Gateway Deployment of Istio was hardcoded. This for once
would require to use the default label in the examples used by
Istio and not a custom set. On the other hand it would not allow to
use two different Ingress Gateways to serve multiple GerritClusters
in the same cluster, which is a requirement, if using SSH, since
Istio can't route TCP requests based on the hostname.

This change allows to configure the labels used for selection. The
default stays as is.

Change-Id: I52789b4a0c6e82742e87c9c7c84fc0a8e8e7a256