Allow to use /a/ prefix on push to apache-git-http-backend
The replication plugin will try to fetch, if replication of all refs
is requested. This fetch will usually need to be authenticated, since
it is routed to Gerrit. Thus, the URL used for replication needs the
/a/ prefix. However, the apache-git-http-backend couldn't use it
for the git-backend, since it would assume that /a/ is part of the
repo path.

Now /a/ was added as a ScriptAlias, allowing to use it as a prefix.

Change-Id: I625aa610416978c4b1d4625ce5dc5a89f6d8a84d