[Operator] Also reconcile dependent resources on additions
The Java Operator SDK by default does not trigger a reconcile on
dependencies, if the primary resource was modified with only an
addition to the spec. This is an optimization to reduce unnecessary
reconciliations, since for example ValidatingWebhooks can add default
values to a resource, which do not change the reconcile result.
However, this optimization does not work, if the resource uses lists.

To avoid this issue, now reconcile is enforced, even if it is only
an addition. Performance-wise this should be fine, since the Gerrit
Operator typically will not manage hundreds of resources, like an
operator for a stateless microservice based application.

Change-Id: I1a5c35798f94510685c948472ce17e1c7a9c26ab