Add option to preserve packs during git-gc
The custom command git-gc-preserve hard-links the old packs to
objects/pack/preserved/ before git gc is executed. This helps JGit
to avoid races between git gc trying to prune old packs and another
concurrently running git command (most often fetch) which may fail to
access a pruned pack file. If JGit doesn't find some object it will
check if it can be found in objects/pack/preserved and if it finds it
there will hard-link it to objects/pack to ensure the next run of
git-gc-preserve will not prune it.

Change-Id: I88029a1c4bb7ba8d074dd5e8fdeed04f21d4560f