[Operator] Change process to create new CRD version
To create a new CRD version, the Java code representing the CRDs
had to be fully copied. That led to huge changes that were hard to
review and to a lot of duplicated code.

This change changes the procedure. The final CRD file containing
both versions is now generated by a script during the build based
on the checked in previous CRD version.

To create a new version, run `operator/tools/newCRDVersion $NEW_VER`.
This will update the version name in the Java code, which is now
maintained in only a single place, and all references in yaml files.
Then the script will delete the obsolete version and mark the
previous version as deprecated.

The script will also trigger the build of the operator, which will
create CRD yaml files that only contain the new version. The build
will then copy the new CRD yaml files to the `./crd/current` directory
to be checked in and then run the `./operator/tools/combineCRDVersions`
script, which will create a single CRD yaml file containing both
the previous and the new version and store it in the
gerrit-operator-crds template directory.

The developer can than do their changes to the API and by rerunning
the maven build, the CRD yaml files will be updated accordingly.

Change-Id: Ibf1691b1a5ec85bc68adb8e586d1146ea9395aef