[Operator] Add Fluent Bit Sidecar for Logging
Each Gerrit pod can have a fluent bit sidecar attached.

By default the sidecar is configured to simply tail Gerrit log files,
tag log messages with the file name they are sourced from, and output
to stdout. New configuration can be provided to the sidecar as a string.

The operator sets up the input configuration as that depends on the
log volume.

Configuration options can be found:
https://docs.fluentbit.io/manual/administration/configuring-fluent-bit/classic-mode/configuration-file

Change-Id: I468c6b4b18848a07cfc33b1a53d3da3ef5d6a31e