Implement feature toggle to allow Gerrit Multisite deployment
This commit introduces a feature toggle to enable the deployment of
Gerrit Multisite. It is defined as an environment variable called
`MULTI_SITE`, disabled by default to preserve current behaviour.

However, enabling this feature at the moment triggers a
java.lang.UnsupportedOperationException exception as the implementation
for Multisite is not yet provided.

Change-Id: I3a7652556fbd3d6b4fe6154a7ae0279180a0fad7