Implement a cluster mode switch to allow Gerrit Multisite deployment
This commit introduces an enumeration to enable the deployment of
Gerrit in Multisite or High-Availability mode.
It is defined as an environment variable called `CLUSTER_MODE`,
set to `HIGH_AVAILABILITY` by default to preserve current behaviour.

However, setting the mode to `MULTISITE` at the moment triggers a
java.lang.UnsupportedOperationException exception as the implementation
for Multisite is not yet provided.

Change-Id: I3a7652556fbd3d6b4fe6154a7ae0279180a0fad7