Introduce feature toggle for disabling the shared file system
This change is the first in a chain that will allow to provision
Gerrit HA with or without a shared file system. This change does
not provide any implementation in the case that of disabling the
shared file system and it throws
java.lang.UnsupportedOperationException exception.

It's worth noting the introduction of a new feature flag called
`sharedFileSystem` in the gerrit-operator Helm chart. By default,
its value is set to `true`, preserving the existing behaviour and
ensuring backward compatibility. However, when set to `false`, the
operator permits to provision a Gerrit HA installation without
sharing the file system for git and logs data.

Change-Id: I3a7652556fbd3d6b4fe6154a7ae0279180a0fad7