Implement feature toggle for activating Gerrit Multisite deployment
This commit introduces a feature toggle to enable the deployment of
Gerrit Multisite. It adds a property named `multisite` to the Helm
chart configuration, which defaults to false to maintain backward
compatibility. Enabling this feature allows the Gerrit operator to
provision a Multisite installation. However, note that currently,
enabling Multisite triggers a java.lang.UnsupportedOperationException
exception as the implementation for Multisite is not yet provided.

Change-Id: I3a7652556fbd3d6b4fe6154a7ae0279180a0fad7