Improve deletion of empty directories under refs/changes
The deletion of empty directories was implemented in the way that it
tried to delete *every* directory relying on the fact that the rmdir
command will fail for non-empty directories. In addition to that the
deletion was done in two steps because refs/changes refs have two levels
of directories:

  refs/changes/45/12345/

In the first iteration the "12345" would get deleted (if empty) and in
the second iteration the "45" directory would get deleted (if empty).

Improve by relying on the find command to perform deletion:

  find refs/changes -type -d -empty -delete

This finds all empty directories and deletes them. The traversal is done
in the depth-first order which means that the empty directories which
are the leaves of the directory tree get deleted first. This means that
one execution of that command is enough to delete all empty directories
including those that become empty because of the deletion of their empty
subdirectories.

Change-Id: Ie139176eb5d81ac962dd451345c8c270edbe25a0