Provision a gerrit cluster without nfs
This change allows the k8s-gerrit operator to deal with multiple
replicas in the same stateful set with or without having a shared
file system based on the feature flag introduced in change 409743.

When the cluster mode is set to `MULTISITE`, each K8s replica
will have its own persistent volume to store the gerrit site.

Change-Id: Ia07c495ef1d3e3dad349a333e05625f6bea73487