Provision a gerrit cluster without nfs
This change allows the k8s-gerrit operator to deal with multiple
replicas in the same stateful set with or without having a shared
file system based on the feature flag introduced in change 1176319.

It creates a StatefulSet where each replica has its own
PersistentVolumeClaim to store git data and logs. To achieve this,
the k8s-gerrit operator has been updated:

- Add a new reconciler GerritClusterNoSharedFSReconciler
- Bind different reconcilers levaraging the feature flag
- Update Gerrit custom resource model to configure the PersistentVolumes
  using a new VolumeClaimTemplate
- Add an headless service for internal DNS resolution

Change-Id: Ia07c495ef1d3e3dad349a333e05625f6bea73487