Expand gerrit-init to set up pull replication
When using pull replication gerrit-init container
is setting up the remotes for pull replication.

To do this a new step in the initializer has been
added that:

- check if pull-replication needs to be used
- create the specific configuration for each
  pod in the statefulset

The pod identifier in the statefulset is derived
from the `pod-name` already available in the operator.

Please note that the configmaps are immutable so
the PullReplicationConfigurator reads the common options
specified for the remotes and use them to populate each
remote section in replication.config.

Change-Id: I8bbb4ccd47ce2a65eff5bf6c59b655e0ae53d860