Expand gerrit-init to set up pull replication
When using pull replication gerrit-init container
is setting up the remotes for pull replication.

To do this a new step in the initializer has been
added that:

- check if pull-replication needs to be used
- create the specific configuration for each
  pod in the statefulset leveraging the pod-index label

Please note that the configmaps are immutable so
the PullReplicationConfigurator copies them in the
filesystem and use them as a template for setting
replication.config.

At the moment PullReplicationConfigurator does not
support auto scaling.

Change-Id: I8bbb4ccd47ce2a65eff5bf6c59b655e0ae53d860