Implement Istio traffic management configuration
This commit introduces a new reconciler,
GerritIstioNoSharedFSReconciler.java,
which is responsible for configuring traffic [1] through the lifecycle
management of Custom Resources: Gateway [2], VirtualService [3], and
DestinationRule [4].

Inbound traffic configurations are specified as follows:
- HTTP protocol is allowed on port 80.
- HTTPS protocol is allowed on port 443 (configurable; can be disabled).
- TCP protocol is allowed on port 29418 (configurable; can be disabled).

Session affinity [5] is established based on the source IP address,
ensuring that subsequent requests from the same client IP are
consistently routed to the same backend server. This enhances session
persistence and maintains a continuous user experience.

References:

[1] https://istio.io/latest/docs/concepts/traffic-management/
[2] https://istio.io/latest/docs/reference/config/networking/gateway/
[3] https://istio.io/latest/docs/reference/config/networking/virtual-service/
[4] https://istio.io/latest/docs/reference/config/networking/destination-rule/
[5] https://istio.io/latest/docs/reference/config/networking/destination-rule/#LoadBalancerSettings-ConsistentHashLB

Change-Id: Ib24992d2d2b6b4457e729c7f18c62ad8af59adab