[Operator] Fix restart of Gerrit on Secret update
The automated restart of Gerrit pods, if the Secret containing
secret Gerrit configuration was updated did not work properly
anymore.

This change changes the way to handle events on the Secret back
from using a read-only dependent resource to using an event source.
This matches what is used for module secrets.

Change-Id: Ie8d6c5a37ca64018e363232ae261d81672678f69