Update java-operator-sdk to 4.4.4
This update includes an update of the fabric8 kubernetes client
to 6.7.2.

The new SSA based way of installing dependent resources was disabled
for now to avoid potential migration issues. This should be enabled
and tested separately.

JOSDK 4.4.4 has a bug, where equality matching of ConfigMaps and
Secrets does not work as expected. This breaks the feature to restart
Gerrit pods on a config change. This bug has been fixed in later
versions, which will be applied in follow up changes.

Change-Id: Ibaaa45d43928f6b7fa383cfacb7ed04c917f7057