[Operator] Customize fluentbit multiline parser
Previously, the default fluentbit java multiline parser separated
starting lines of stack traces. This replaces the default parser
to display stack traces as one log entry.

The logic is: If a line starts with square brackets which
indicate a timestamp it is considered a new log line. Otherwise
it continues the previous log entry. Logs which are not caught by
the regex will still be sent on as output, but not parsed as
multiline logs.

Requirements and configuration details:
https://docs.fluentbit.io/manual/pipeline/parsers/configuring-parser#parsers-configuration-file
Change-Id: Ib9c170d7e306ccc008f88b01f871d7c577aa5720

Change-Id: I62e81f22a9bdf6ca5a63ef9bdd292fd5f6ca3c2d