Don't store logs on shared volume
Thee logs were stored on a NFS volume. This was a historic artifact,
when logs still needed to be accessed by multiple pods. Since
shared storage is usually more expensive and harder to maintain,
the logs should rather be stored on block storage.

Now, the logs will be stored as follows:
- Gerrit: in the PV containing the Gerrit Site
- Git GC: No persisted logs. Logs will be posted to stdout.
- Receiver: No persisted logs. Logs will be posted to stdout.

BREAKING CHANGE: This will break any setup using the Gerrit Operator
and relying on logs being stored in the NFS volume.

Change-Id: If17368d0bee1769666c05daa3282cbb6572a156e