Don't align values of FluentBit config
A typical style used in FluentBit configuration is to align the values
into a single column using multiple whitespaces. However, this was
not done consistently here. Also this style is hard to maintain, since
the number of required whitespaces might change with every new option.

This change reformats the configuration to use only a single whitespace
between key and value.

Change-Id: I2a3179b36df6f4fc84175eedc9f7f8d3b660e59e