Store current apiVersion as annotation in dependent resources
If the apiVersion was updated, it could happen that the apiVersion
of dependent resources was not updated in ETCD. This happened, if
the spec in the dependent CustomResource was not updated. In that
case the next update would fail, if the original apiVersion was not
present in the CRD spec anymore.

The operator will now store the apiVersion of dependent CustomResources
as an annotation in them. On reconcile if that annotation doesn't
match anymore, the complete resource will be reconciled and the
apiVersion updated.

This change will add a few potential reconciles, since Kubernetes
can also set annotations like
kubectl.kubernetes.io/last-applied-configuration. This would now also
trigger a reconcile, but that should be negligible.

Change-Id: I130a486917d402a68d145ff94d2e1f5e20c80f05