[Operator] Use host IP based hashes for session stickiness in istio
So far the Gerrit Operator configured Istio to use cookies to determine
session stickiness. These cookies had a TTL of 60s. This caused some
usability issues in the Gerrit UI, when writing comments or editing
a change. In this cases, if the cookie expired and the user's request
were sent to the other instance, it could happen that the data on
this second instance had not caught up, e.g. due to NFS latency. In
that case it could happen that the user's input was reset in the UI.

Now, Istio is using the host IP for session stickiness. This is more
likely to cause imbalanced load, since CI systems who typically have
a stable IP will stick to a single instance and CI systems might thus
by chance mostly end up on a single instance and cause a high load.
This should be observed. However, the advantage over just using a
cookie with a longer TTL is that this way also supports SSH, REST
requests done from outside the browser and git, which should overall
lead to a more stable experience.

Since the NGINX ingress provider does not support this feature, it
will continue to use cookies, but with a TTL of 1h.

Change-Id: I4714e9252a4fa7aa9532d3a66e7d6c49f4017a12