Add option to provide executable scripts for plugins
Some plugins like the hooks plugin use data files that have to be
executable. Kubernetes by default uses the 0644 file mode to mount
files.

This change allows to set an option per plugin to make mounted files
have the file mode 0754 instead, making them executable.

A better way to do this would be to allow to have multiple secrets
or configmaps and to make the executable option available per
configmap or secret. That way not all data files of a plugin would
become executable. This would however be a breaking API change
requiring a conversion webhook. Conversion webhooks are as of now
not supported. Thus, to enable this feature faster, this change
is provided as is and the API will be adapted at a later point in
time. Since the configuration is in the hands of the administrator
only, the risk of having unnecessary or even malicious executable
files should be minimal/not exist.

Change-Id: Ic52b7e42517fbb109184f2ac71e856afb1f61cd3