Fix inter-Gerrit communication using Istio
A bug in Envoy keeps connections open even if the IP address the target
hostname resolves to changed, which happens during rolling restarts [1].
This meant that the communication between Gerrit containers using the high-
availability plugin was only working in one direction almost all the time.

This change adds ServiceEntries for the primary Gerrit pods. This works
around the issue.

[1] https://github.com/envoyproxy/envoy/issues/19458

Change-Id: I899eca647a7294e0deb877ee51533d6147e1d9d6