[Operator] Create VolumeMounts for shared Volume in dedicated factory
The classes describing the CustomResources also provided
utility methods to create parts of dependent resources.
This was a bad design, disregarding separation of concerns.
Creating parts for the dependent resources should be the
concern of the reconcilers.

Due to the above named issues, the API classes could not
be moved into a separate module.

Change-Id: I276e1addb10bbc679963e73e224cc569df305b5e