Add GerritCluster/Gerrit CR webhook validation for Multi-site mode
This changes introduces validation for Multi-site mode in the webhook
when GerritCluster or Gerrit CR are applied in the cluster.

For GerritCluster CR the validation checks that there is only one Gerrit
configuration and it must be Primary.

For Gerrit CR the validation checks that refdb configuration must be
provided.

Change-Id: I5c0bf1dddec3a2bcf469b1e1387040f72894b928