Add GerritCluster/Gerrit CR webhook validation for Multi-site mode
This changes introduces validation for Multi-site mode in the webhook
when GerritCluster or Gerrit CR are applied in the cluster.

For GerritCluster CR the validation checks that there is only one Gerrit
Primary and more than one replicas.

For Gerrit CR the validation checks that refdb and events broker
configuration must be provided.

Change-Id: I930dd5cf749180cc6de6b81b7a5102d21becc4fd