Bazel: Fix running tests from JAR archive
Move resources to the test directoris, so that it can be found from the
JAR archive. Maven build is copying the resources to the target
directory and that's why it's found with the relative path. For the
places where the resources are using for different java packages,
normalize the resource name and load it with complete package name.

Also rename the abstract base classes with *Test suffix to *TestBase
class to avoid from loading them as regular tests in Bazel build.

Change-Id: Ibde60898661e17fce90eab3c549e0613137c6a08