Add option to copy created indexes to specified output dir
This change adds an option to the entrypoint script of the gerrit-indexer
container to specify an output directory. The script will copy the resulting
indexes to the specified directory, which can be a mounted volume. This allows
to move the indexes to a Gerrit instance e.g. by using a RWM-mode volume.

Change-Id: I535572b2b90d8d7aaea9cc13f0cebca623671ded