[Operator] Add CustomResource to create GerritIndexer job
By default the indexes used by Gerrit will be updated using online
reindexing. On large sites this might be quite costly and affect
performance in the primary Gerrit instances serving clients.

To avoid performance issues, this change adds a new CustomResource that
manages a separate Gerrit instance running a s a job and not exposed
to clients that will run offline reindexing on a copy of the repositories.
Afterwards it will copy the resulting indexes to a given output volume
which in the followup can be mounted to the Gerrit instances to apply
the new index versions.

Change-Id: I1c1476bab82b992b259e3e545763e0a43759ac31