Don't reference OperatorContext in API module
The API module should purely describe the CustomResources and not care
about the state of the operator.

This change moves the affected condition to the callers. This is not an
ideal but the most similar solution to the current state. A future change
should investigate the use of different reconciler implementations to
reduce if statements.

Change-Id: Ib87bb939ef535296c915eac8283172e5bb0861ba