Build container images for Gerrit deployment using Bazel
So far container images used in the Kubernetes deployment of
Gerrit were built using Docker and some shell scripts. This
way of building the container images did not provide 100%
reproducible results, i.e. they didn't follow OCI standards.

This change introduces a Bazel build for the container images
used in the Kubernetes deployment of Gerrit.

The resulting container images are 99% reproducible. The only
factor that might lead to changed images are the Gerrit plugins
for which no stable URLs exist. Thus, the plugin files may
change. This should be fixed in the future.

Apko [1] is being used to provide reproducible alpine-based
base container images.

There is no way to on the fly change the Gerrit executable
or plugin jar files anymore.

The old build files have been removed. The test suite for the
container images was not yet adapted. This will happen in a
followup change.

[1] https://github.com/chainguard-dev/apko

Change-Id: I2245b9670a6c6524c0f93cf8cc918bc4b8580665