Delete previously shared index before providing new iteration
If the GerritIndexer was run twice for the same index version, it
would not delete the previously published index. Instead it would
nest the new copy within the previously published copy and from then
on override everything, which could lead to an unpredictable state.

Now, the target directory is being deleted before copying the index.

Change-Id: Ib281bd951b9e85fd11a683370d829bf2365780a9