Use master branch for Gerrit executable and plugins
The k8sgerrit setup in part relies on features explicitly added
to Gerrit to support features required for running it in Kubernetes.
If we would continue to use releases, we would have to wait for
up to half a year to be able to use those features.

For supporting stable releases, the k8sgerrit project currently
lacks the manpower. This might happen in the future.

Change-Id: Idedfb37cdc5ccfa14fbad7c723d6ac1a0b59bed8