Make more use of constants in gerrit-initializer
The gerrit-initializer is very closely tailored to work with the setup
as provided with the k8sgerrit setup. However, there were some values
that were implemented as variables instead of constants. This didn't
work well together and in the end didn't add a lot of value.

Now more constants are being used, simplifying the gerrit-initializer.

Change-Id: Iff2ef6fe1c81afd9f1665c4d5321cca51fcf5efc