Add SecurityContext to containers
This adds a SecurityContext to almost all containers managed by the
Gerrit Operator. The only exception is the NFS workaround InitContainer,
because it needs to run as root. The risk is however low, since it is
only running for a very short time, if at all.

The new SecurityContext sets the user ID and group ID as well as not
allowing to run as root.

Change-Id: I45bc0966b698ea77326391358cf0576b1b678aa9