Add Gerrit-Maintenance CR for managing maintenance tasks
A Gerrit site requires to run a number of maintenance tasks
regularly. The strategy so far was to have a separate CR per
task. This would not scale with the growing number of
supported tasks.

This change adds the GerritMaintenance CR that is managed
by the GerritCluster resource. This resource is meant to
provide support for maintenance tasks in a Gerrit site.
At the moment, it only supports Git GC, but further tasks
will be added in future changes.

The GitGarbageCollection CR still exists, but is now
deprecated.

Change-Id: I41fc7fb5d56de0c9618a3b8c1dacab86bae42c72