Improve detection of API version update
The operator needs to trigger a reconcile, if the API version
of a dependent resource happens. However, such an update is
invisible to the operator, if nothing but the version was updated.

To solve this, the operator adds an annotation storing the applied
version. This didn't work perfectly so far.

Now, the annotation handling is integrated better with the operator
procedures.

Change-Id: Id0cb6ad9a0314d43cfac138b3cd952e291717ac0