Mount EmptyDir volume to "/tmp" dir in Gerrit containers
Some Gerrit functionality relies on creating temporary files
under `/tmp`, which was not working due to the recently
introduced read-only filesystem.

This change mounts the 'tmp' subdir in the 'tmp'-emptyDir
volume to '/tmp'. To avoid namespace clashes, the mounted
'/home/jenkins' dir is now using the 'home' subDir. This
should not be a breaking change, since these contain only
temporary files.

Change-Id: If9141aae9a26e9d4b5a2a0d88e44f7292bef20bb