Enforce Java tmp dir to be on the site volume
The recent H2 update in Gerrit caused large temporary files
to be written to /tmp. These would lead to exceeding of the
max size of the emptyDir volume used for /tmp and lead to a
pod restart.

To avoid this the Java tmp dir is now under /var/gerrit/tmp/java,
which is on the persistent volume containing the Gerrit site,
which should be large enough to handle these files.

Change-Id: I9066bff717cb1c8549370fd847a79974e68713ff