Mount EmptyDir volume to "/tmp" dir in Indexer containers
The JVM relies on creating temporary files under `/tmp`, e.g. jstack
was not working due to the recently introduced read-only filesystem.

This change mounts the 'tmp' subdir in the 'tmp'-emptyDir volume to
'/tmp'. To avoid namespace clashes, the mounted '/home/gerrit' dir is
now using the 'home' subDir. This should not be a breaking change, since
these contain only temporary files.

Change-Id: Id550360c442f8ad8e745fe5cb1d2b90758825636