In pre-stop hook first send SIGTERM and then create thread dump
This ensures we first trigger graceful shutdown of the Gerrit
application before triggering a thread dump which might fail if
available system resources are low. Ignore the return code of the
involved commands to prevent that a failing jstack would skip running
the tail command which would have the effect that kubernetes would
forcefully shutdown the pod.

Change-Id: I63da60cabed9b971cd936ceb6d99b73d93c4598e