Remove unmaintained plugins from gerrit-base image
Some plugins related to the multisite setup are currently not
maintained in the Opensource due to the move to a BSL license of
Gerritforge plugins. Due to that they are not available on the Gerrit
CI and the container images fail to build.

Remove those plugins from the container. The multisite setup can still
be used with either custom container images or by just configure those
plugins to be downloaded on startup.

Change-Id: I5d9f1c53ea06e924b869682e409de2c52d786760